Fix compilation issue on OSX By changing the type of last_item_idx from size_t to uint64_t, a compilation issue on OSX is fixed. Those types should be equivalent on most modern machines anyway

+++ Archipelago.cpp
@@ -67,7 +67,7 @@ void (*setreplyfunc)(AP_SetReply) = nullptr;
 // Serverdata Management
 std::map<std::string,AP_DataType> map_serverdata_typemanage;
 AP_GetServerDataRequest resync_serverdata_request;
-size_t last_item_idx = 0;
+uint64_t last_item_idx = 0;
 
 // Singleplayer Seed Info
 std::string sp_save_path;
